Infer lifted combinator reductions automatically
Combinators are now defined as proper constants, which makes it easier
to use them in internal proofs. The names of these constants are
typically used as keys for objects which refer to specific combinators.
In general, the set of supported combinators has become more flexible.
The bracket abstraction algorithm is still based on fixed rules, though.

When registering a new applicative functor, the lifted combinator
reductions proven by the user are completed automatically, according to
an extendible set of rules (attribute `combinator_eq`). If the
composition and/or identity laws are missing, they are added implicitly
to the proof obligations. The user-facing names of the laws are
determined by the invocation of `Applicative.declare_combinators`. The
theorems declared as `combinator_repr` control how combinators are
presented to the user in theorems and proofs. The proof state created by
the `applicative` command has changed slightly, in particular the order
in which type variables are invented.

Remaining work:
  - fix bug with extra type variables on rhs of combinator equations
  - warn user if proven reductions are redundant (requires separate
    book-keeping of derived combinators)
  - handle interchange law properly
  - note theorems in theory (again)
  - complete list of combinator derivations
  - improve user interface, especially display of combinators
  - use S combinator in bracket abstraction