Support automation of interchange law
The automatic combinator inference now supports the interchange
combinator T'. This means that T' can be inferred from other
combinators, and that it may be used in further derivations.

The declaration of combinators has been extended with a list of argument
positions. These refer to arguments of the combinator which should be
lifted as pure terms. This way, the first argument of T' is restricted
to a pure value in the lifted reduction.

The user-facing proof obligations present the interchange law in its
usual formulation. This requires some extra steps, because the
transformation between this form and the lifted reduction of T' needs
the homomorphism law.