Distinguish names of noted theorems
There was a collision with the datatype package (<functor>.rel_intros),
which caused an error in the Stern_Brocot theory. Some of the new names
are taken from an older version of note_afun_thms.
pure_comp_conv is no longer named, because its use is mainly internal to
the package. If necessary, it can be obtained from pure_B_conv via
instantiation.